+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-26 15:23:31
+
+### Changed Files
+- `src/services/analytics/policyStatusService.ts`
+- `vite.config.ts`
+
+### Statistics
+```
+ 2 files changed, 51 insertions(+), 8 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-26 14:29:20
 
 ### Changed Files